only increase density in high value sections
