+++ src/insert/spacefinder/utils.ts
@@ -1,0 +1,18 @@
+const highValueSections = [
+	'business',
+	'environment',
+	'music',
+	'money',
+	'artanddesign',
+	'science',
+	'stage',
+	'travel',
+	'wellness',
+	'games',
+];
+
+const isInHighValueSection = highValueSections.includes(
+	window.guardian.config.page.section,
+);
+
+export { isInHighValueSection };
